some bug fixes read below
1) fixed scopes and other ammos not dropping from player if the player dies
2) fixed scopes stuff completely disappearing after dying, and doesnt reappear, due to a wrong class name provided in document.getElementsByClassName.
3) Made cleanups actions after disconnection aswell, to be a failsafe if the server disconnects instead.
4) Fixed ammo counter still showing ammos after dying.
5) changed ammos ammo count from Int8 to Int16 for compatiblity of larger ammo drops, although now at the cost of an extra byte required :(

+++ client/src/deserialisers.ts
@@ -196,7 +196,7 @@ export function deserialiseMinEntities(stream: IslandrBitStream) {
         }
         else if (type == EntityTypes.AMMO) {
             const minEntity = Object.assign(baseMinEntity, {
-                amount: stream.readInt8(), color: stream.readInt8(),
+                amount: stream.readInt16(), color: stream.readInt8(),
                 hitbox: <MinHitbox>{
                     type: "rect", width: 2, height: 2,
                 } })
